altered tasks smallint columns to integer because mysql is stupid

+++ database/migrations/2016_04_08_064233_create_tasks_table.php
@@ -16,8 +16,8 @@ class CreateTasksTable extends Migration
             $table->increments('id');
             //$table->integer('user_id')->unsigned();
             $table->string('title');
-            $table->smallInteger('priority')->default(1);
-            $table->smallInteger('completed')->default(0);
+            $table->integer('priority')->default(1);
+            $table->integer('completed')->default(0);
             $table->timestamps();
             // Foreign Keys
             // $table->foreign('user_id')
